fixed most bugs found by Fingbugs. Only two categories of bugs are not fixed (execute sql from nonconstant string, and use of float instead of double).
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/ReFlow/trunk@10831 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Controller/src/java/org/gusdb/workflow/Workflow.java
@@ -413,7 +413,7 @@ public class Workflow<T extends WorkflowStep> {
                 Integer ws_id = rs.getInt("workflow_step_id");
                 String stat = rs.getString(state_str);
                 Date endTime = rs.getTimestamp("end_time");
-                float spent = rs.getFloat("hours");
+                double spent = rs.getFloat("hours");
 
                 if (oneColumnOutput) System.out.println(nm);
                 else {
@@ -421,7 +421,7 @@ public class Workflow<T extends WorkflowStep> {
                     formatter = new Formatter(sb);
                     if (spent != -1) {
                         int hour = (int) Math.floor(spent);
-                        int minute = Math.round((spent - hour) * 60);
+                        int minute = (int)Math.round((spent - hour) * 60);
                         if (endTime != null) {
                             formatter.format("%6$tm/%6$td/%6$ty %6$tH:%6$tM:%6$tS %1$03d:%2$02d %3$-8s %4$-12s  %5$s",
                                 hour, minute, stat, ws_id, nm, endTime);
